% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agedotoliths.R
\name{agedotoliths}
\alias{agedotoliths}
\title{Summary of NORPAC Otoliths by Year}
\usage{
agedotoliths(agedata)
}
\arguments{
\item{agedata}{A data frame of NORPAC ages, often called atsea.ages.}
}
\value{
A data frame of proportions by year and month. The 
function also saves a summary file (csv format) to the disk
with sample sizes from which proportions can be calculated.
}
\description{
Create a proportion table of age samples that have been aged
for the NORPAC data. This information is normally given to the
stakeholders at the December JTC meeting.
}
\author{
Kelli Faye Johnson
}
