% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.pctages.R
\name{as.pctages}
\alias{as.pctages}
\title{Converts the rows of a matrix to percentages cleanly as.pctages}
\usage{
as.pctages(x)
}
\arguments{
\item{x}{A matrix}
}
\description{
Converts the rows of a matrix to percentages cleanly as.pctages
}
