% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AtSeaComps.R
\name{atseacomps}
\alias{atseacomps}
\title{Workup At-Sea Composition Data for Pacific Hake}
\usage{
atseacomps(
  atsea.ages = NULL,
  ncatch = NULL,
  endyear = hakedata_year(),
  ages = 1:15
)
}
\arguments{
\item{atsea.ages}{An R object with NORPAC ages. The default value is
\code{NULL}, where the object will be read from a saved object within
a directory called extractedData.}

\item{ncatch}{An R object of NORPAC catches.
The default value is loaded from the disk using \code{\link[=loadlocal]{loadlocal()}}.
Otherwise, the R object is typically stored in the \code{hakedata} environment.}

\item{endyear}{The final year of the data, where the data will start
in 2008.}

\item{ages}{A vector of ages to be included in the composition data.
The default is to include ages from one to fifteen.}
}
\value{
A list object with summary information regrading the composition
data. Multiple files are saved to the disk as well. These saved files
include summary information in ...report.txt files and comps.csv files.
}
\description{
todo: add more information about this function here.
}
\author{
Kelli Faye Johnson
}
