% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catchpullcompare.R
\name{catchpullcompare}
\alias{catchpullcompare}
\title{Compare NORPAC and PacFIN At-Sea Catches}
\usage{
catchpullcompare(pcatchatsea = NULL)
}
\arguments{
\item{pcatchatsea}{A data frame of catches from PacFIN of the 
at-sea sectors by year and month. If \code{NULL}, then the function 
will import the \code{.Rdat} file.}
}
\value{
A data frame of comparisons between the at-sea catches generated
from NORAPC and PacFIN. The "diff" column are the differences by month 
and year between the two data bases.
}
\description{
The hake stock assessment uses NORPAC catches as data, but it is a
useful exercise to ensure that the catches from NORPAC match those
coming from PacFIN. The databases should match within 3 days according
to Vanessa Tuttle.
}
\author{
Kelli Faye Johnson
}
