% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commit-.R
\name{commit_catch}
\alias{commit_catch}
\title{Updates the most recent catch time series in the SS3 data file}
\usage{
commit_catch(
  dir_data = hakedata_wd(),
  digits = 5,
  lastassessmentvals = c(NA, NA, NA)
)
}
\arguments{
\item{dir_data}{A directory that houses the \code{.csv} data files for the stock
assessment. These are typically found within the \code{data} directory of the
hake-assessment repository.}

\item{digits}{The number of digits you want the results to
saved to. \verb{[round](..., digits = digits)} will be used.}

\item{lastassessmentvals}{A vector of three values,
\itemize{
\item estimate of spawning stock biomass for terminal year in thousands of
metric tons,
\item estimate of depletion in percent with one decimal place,
\item estimate of TAC for the first year of the forecast period in metric tons
rounded to the nearest whole number.
}}
}
\value{
The following two files are updated in the
pacific-hake/hake-assessment repository:
\itemize{
\item landings-tac-history.csv
\item catch-targets-biomass.csv
}
}
\description{
Integrate the latest catch information from the U.S. and Canadian Pacific
Hake fisheries into the Stock Synthesis data file. Information on catches is
found in \code{.csv} files saved in the hake-assessment repository.
}
\author{
Kelli F. Johnson
}
