% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertToDecDeg.R
\name{convertToDecDeg}
\alias{convertToDecDeg}
\title{Convert to Decimal Degrees}
\usage{
convertToDecDeg(x, rnd = 2)
}
\arguments{
\item{x}{A vector of integer values, where the last two digits are assumed
to be the minute values that will be transformed into a decimal value.}

\item{rnd}{An integer value specifying how many digits to round to using
the argument \code{digits = rnd} in \code{round}.}
}
\value{
A vector of numeric values.
}
\description{
Convert a four- or five-digit number to decimal degrees, where it is assumed
that the last two digits of the input data are minutes.
}
