% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wtatage_calculations.R
\name{data_wtatage}
\alias{data_wtatage}
\title{Create weight-at-age files for hake assessment}
\usage{
data_wtatage(
  dir,
  maxage = 15,
  yrs = 2008:hakedata_year(),
  navgyears = c(5, 5),
  nforecast = 3,
  maturity = c(0, 0, 0.261, 0.839, 0.961, 0.92, 0.928, 0.926, 0.957, 0.944, 0.98, 0.962,
    1, 0.958, 0.955, 0.9, 0.9, 0.9, 0.9, 0.9, 0.9)
)
}
\arguments{
\item{dir}{The directory where the data is stored.
It can either be relative or absolute because the working
directory will not be changed. Instead, \code{dir} is
used to import data and export results to.}

\item{maxage}{The age of the plus group used for the stock assessment.
This will correspond to the maximum age group in the data, not in the
model because SS can model many ages when there is only information in
the data for a few ages.}

\item{yrs}{A vector of years to search for recent data. Typically,
the vector starts with 2008 and ends with the most recent year
of data. This will allow files created from \code{wtatage_collate()} to
be included in the analysis, i.e., recent US data. Typically, you
should not have to change this value from the default entry.}

\item{navgyears}{The number of early and late years to average since
1975 and \code{max(yrs)} for the early and late analysis asked for
by the Scientific Review Group in 2017. The argument can be a single
value or a vector of two values, where in the latter case the second
value will be used for the most recent time period.}

\item{nforecast}{The number of years to forecast into the future.
Typically, this is three for the hake assessment and will lead to
this many rows of mean weight-at-age data being copied to the data frame
beyond the last year of empirical data.}

\item{maturity}{A vector of maturity values from the maturity ogive.
The length needs to be the same as the number of ages in the model,
not the number of ages in the data.}
}
\value{
todo: document return
}
\description{
Create weight-at-age files for the hake stock assessment.
Updated csv or rds files must exist prior to running.
}
\author{
Ian Taylor
}
