% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatoassessment_catch.R
\name{datatoassessment_catch}
\alias{datatoassessment_catch}
\title{Copy Catch Files to Assessment}
\usage{
datatoassessment_catch(dirout)
}
\arguments{
\item{dirout}{A directory that houses the \code{.csv} data files
for the stock assessment.}
}
\description{
Copy the catch files from the US Catches directory to the assessment
data directory.
}
\author{
Kelli Faye Johnson
}
