% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatoassessment_comp.R
\name{datatoassessment_comp}
\alias{datatoassessment_comp}
\title{Copy U.S. composition data to the hake-assessment data directory}
\usage{
datatoassessment_comp(dirout)
}
\arguments{
\item{dirout}{A directory that houses the \code{.csv} data files
for the stock assessment.}
}
\description{
Copy U.S. composition data to the hake-assessment data directory
}
\details{
TODO: Write these files out directly to hake-assessment rather than writing
to a temporary file. Make sure that no code depends on the temporary
files before doing this. Eliminate this function. Write documentation
in the function that makes these files that the original files in
hake-assessment are overwritten.
TODO: Make a message that encourages a commit in hake-assessment to update
the data.
}
