% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.matrix.R
\name{extract.matrix}
\alias{extract.matrix}
\title{Extract.matrix extracts a matrix from a table object}
\usage{
extract.matrix(x)
}
\arguments{
\item{x}{A table object.}
}
\description{
Extract.matrix extracts a matrix from a table object
}
