% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_big_number.R
\name{format_big_number}
\alias{format_big_number}
\title{Format an object with a comma}
\usage{
format_big_number(x, digits = 0)
}
\arguments{
\item{x}{A vector that will be passed to \code{\link[=format]{format()}}.}

\item{digits}{The number of digits you want to include
when rounding and using \code{round(digits = digits)}.}
}
\value{
Character values are returned using \code{\link[=format]{format()}}.
}
\description{
Format an object using a comma and rounding to the nearest integer.
}
\examples{
# Remove tenth
format_big_number(1.1)
}
