% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sum.R
\name{get_sum}
\alias{get_sum}
\title{Get sum of data without \code{NA} values}
\usage{
get_sum(x)
}
\arguments{
\item{x}{An R object containing numeric values.}
}
\value{
A single numerical value.
}
\description{
Get the sum of all values after excluding \code{NA} entries using
\code{na.rm = TRUE} inside of call to \link{sum}.
}
