% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wtatage_calculations.R
\name{get_wtatagecsv}
\alias{get_wtatagecsv}
\title{Remove Weight-At-Age Outliers}
\usage{
get_wtatagecsv(
  file,
  removeOutliers = TRUE,
  outlierplot = FALSE,
  outlierPlotName = "wt_at_age_outliers.png",
  elimUnsexed = FALSE,
  elimAge = 99,
  verbose = FALSE
)
}
\arguments{
\item{file}{The file path, full or relative, to the \code{.csv} file you
are importing. The file should have the following seven columns:
Source, Weight_kg, Sex, Age_yrs, Length_cm, Month, and Year.}

\item{removeOutliers}{A logical value specifying whether or not outliers
should be removed from the data.}

\item{outlierplot}{A logical value specifying whether or not plots should
be generated. If \code{outlierplot = TRUE} and \code{outlierPlotName = TRUE},
then a text file under a similar name to \code{outlierPlotName} will also be
generated that contains a table of outliers by data source.}

\item{outlierPlotName}{A file path to a \code{.png} file that will be used
to save the plots to the disk. If \code{NULL}, then the plots will be printed
to the screen rather than saved to the disk.}

\item{elimUnsexed}{A logical value specifying if unsexed fish should be
eliminated from the data set.}

\item{elimAge}{A vector of ages to eliminate. The default is to eliminate all
fish with an age of 99.
Use \code{elimAge = NULL} to not eliminate fish of any age.}

\item{verbose}{A logical specifying whether or not to print
information to the screen.
The default is \code{FALSE}.}
}
\value{
A cleaned data frame of weight-at-age information for hake.
The data frame includes the following columns:
\enumerate{
\item Source
\item Weight_kg
\item Sex
\item Age_yrs
\item Length_cm
\item Month
\item Year
\item OutlierL
}.
}
\description{
Get saved weight-at-age data and remove any outliers and/or plot
the data to visualize outliers.
Outliers are screened using two methods.
First, data in September of 2003 are screened for fish less than 0.52 kg
and greater than 45 cm in length.
Second, data are screen for a weight-length relationship beyond that deemed
to be normal, i.e., 20e-6\emph{length^3 and 2e-6}length^3.
Started using this method in 2011.
}
\seealso{
wtatage_collate
}
\author{
Ian Taylor
}
