% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hakedata_prelim.R
\name{hakedata_prelim}
\alias{hakedata_prelim}
\title{Find if date is preliminary data}
\usage{
hakedata_prelim()
}
\value{
A logical.
}
\description{
Find if the current date is prior to the data pull for the assessment or
if the date is after the typical final data pull.
Data should be complete the first week of January.
}
\author{
Kelli F. Johnson
}
