% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hakedata-.R
\name{hakedata_year}
\alias{hakedata_year}
\title{Find last year of data for current assessment}
\usage{
hakedata_year()
}
\value{
The last year of data you want as an integer.
}
\description{
Data is extracted the first Friday in January of the year
following the last year of data. Therefore, if data are extracted
from the databases in January, February, or March, then the terminal
year of data should be the previous year. Else, one would want the
most recent data in the current year as the last year of data included
in the extraction.
}
\author{
Kelli F. Johnson
}
