% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ithweekdate.R
\name{ithweekdate}
\alias{ithweekdate}
\title{Find the date for a weekday}
\usage{
ithweekdate(
  number = 2,
  month = 12,
  year = as.numeric(format(Sys.time(), "\%Y")),
  weekday = "Wednesday"
)
}
\arguments{
\item{number}{The iteration you want, e.g.,
\code{2} for second.}

\item{month}{An integer for the desired month.
Options include values in \code{1:12}.}

\item{year}{The year of interest as a four-digit integer.}

\item{weekday}{The day of the week that you are interested in.
For example, \code{"Wednesday"} is the default.}
}
\description{
Find calendar date for a weekday in a given month.
For example, you might want to know the date for the
second Wednesday of December in the current year.
This is the typical date for the meeting of the
Joint Technical Committee for the Pacific Hake
stock assessment.
}
\examples{
# Find second Wednesday in December of current year
ithweekdate()
# Find third Wednesday in December of current year
ithweekdate(3)
}
\author{
Kelli F. Johnson
}
