% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadlocal.R
\name{loadlocal}
\alias{loadlocal}
\title{Load an object from an \code{.Rdat} file}
\usage{
loadlocal(file, dir = file.path(hakedatawd(), "extractedData"))
}
\arguments{
\item{file}{A file path for either an existing file or
a file that will be written to the disk by the given
function. Paths can be either full or relative.}

\item{dir}{The path to a directory.}
}
\description{
Create a file path from \code{dir} and \code{file} using \code{\link[=file.path]{file.path()}} and
load the object into the local environment while returning it.
\code{loadlocal()} assumes that there is only one object in the \code{.Rdat} file.
The utility of this function comes from being able to quickly load
an R object that was saved as an \code{.Rdat} into a function argument without
having to save it in your work space.
}
\author{
Kelli F. Johnson
}
