% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wtatage_calculations.R
\name{make_wtage_matrix}
\alias{make_wtage_matrix}
\title{Make Weight-At_Age Matrix}
\usage{
make_wtage_matrix(
  dat,
  fleetoption = 1,
  value = "weight",
  months = 1:12,
  getmean = FALSE,
  maxage = 15
)
}
\arguments{
\item{dat}{A data frame created from \code{\link{get_wtatagecsv}} or one that
has the following column names: 
Source, Weight_kg, Sex, Age_yrs, Length_cm, Month, and Year.}

\item{fleetoption}{An integer. 
A value of 1 separates the acoustic-survey data from the rest of the data.
A value of 2 pools all the samples.}

\item{value}{A character value specifying which type of data you are 
interested in. The options include \code{"weight"}, \code{"length"},
and \code{"count"}, where the latter will supply the sample size.}

\item{months}{A vector of months to include in the analysis.}

\item{getmean}{A logical value. If \code{TRUE}, then the first row of the
resulting data frame will be the mean across all years. This mean entry will
have a negative year of 1940.}

\item{maxage}{The age of the plus-group bin. The default is 15 years old.}
}
\value{
A data frame with the first six columns pertaining to metadata, 
i.e., #Yr, seas, gender, GP, bseas, and fleet, and additional columns
pertaining to each age starting with age zero up to the maximum age
supplied with the argument \code{maxage}. Ages for which there were no 
samples are filled with \code{NA}. If \code{value = "count"}, then 
the data frame will return sample sizes for each age rather than mean 
weight- or length-at-age.
}
\description{
Make input file \code{wtatage.ss} for hake.
}
\seealso{
[get_wtatagecsv()]
}
\author{
Ian Taylor
}
