% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wtatage_calculations.R
\name{make_wtatage_plots}
\alias{make_wtatage_plots}
\title{Write Weight-At-Age Plots to Disk}
\usage{
make_wtatage_plots(
  plots = 1:6,
  data,
  counts,
  lengths = NULL,
  dir = getwd(),
  year = format(Sys.Date(), "\%Y"),
  maxage = 15,
  verbose = FALSE
)
}
\arguments{
\item{plots}{A vector of digits that specifies which plots to create.}

\item{data}{A data frame of weight at age data generated from
\code{make_wtage_matrix(..., getmean = TRUE)}.}

\item{counts}{A data frame of weight at age counts generated from
\code{make_wtage_matrix(..., getmean = TRUE, value = "count")}.}

\item{lengths}{A data frame of weight at length generated from
\code{make_wtage_matrix(..., getmean = TRUE, value = "length")}.
The default for this argument is \code{NULL}, and the function can be ran
without supplying length data.}

\item{dir}{The directory that you want to save the plots in. The default
is to save the plots in your current working directory.}

\item{year}{A character value giving the current year that will be used in
the file names for each plot.}

\item{maxage}{The maximum age of fish modelled in the stock assessment, i.e.,
what is the age at which all fish are grouped into a plus group.}

\item{verbose}{A logical value specifying whether or not to print
information to the screen.}
}
\description{
todo: write a description for make_wtatage_plots
}
\author{
Ian Taylor
}
