% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mappingAgeSamples.R
\name{mapfun}
\alias{mapfun}
\title{Plot Map of Fishing Locations for Each Age Sample}
\usage{
mapfun(data, doPNG = FALSE, dir = hakedatawd())
}
\arguments{
\item{data}{A data frame created in mappingagesamples.}

\item{doPNG}{A logical value specifying whether or not to print the
plots to png files.}

\item{dir}{A file path specifying the base directory for which you
have the hake data. Plots will be created in nested folders based on
your input value for this argument.}
}
\value{
Plots or saved png files.
}
\description{
todo: add more details here
}
