% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mappingAgeSamples.R
\name{mappingagesamples}
\alias{mappingagesamples}
\title{Map age samples}
\usage{
mappingagesamples(atsea.ages = NULL, ncatch = NULL, savepng = FALSE)
}
\arguments{
\item{atsea.ages}{An R object with NORPAC ages. The default value is
\code{NULL}, where the object will be read from a saved object within
a directory called extractedData.}

\item{ncatch}{An R object with NORPAC catches. The default value is
\code{NULL}, where the object will be read from a saved object within
a directory called extractedData.}

\item{savepng}{A logical specifying if the png plots of mapfun should
be saved to the disk. If \code{FALSE} the plots will neither be created
or saved.}
}
\value{
A data frame of min and max weights per year.
}
\description{
Map age samples
}
