% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{maturity_at_age}
\alias{maturity_at_age}
\title{Proportion mature at each age}
\format{
\subsection{\code{maturity_at_age}}{

A vector with 20 observations, one for each modelled age starting with age
zero.
}
}
\source{
Melissa Head
}
\usage{
maturity_at_age
}
\description{
A vector of maturity values from the maturity ogive. The length of the vector
is the same as the number of ages in the model, not the number of ages in the
data.
}
\details{
The data are stored in a data object that can be updated/built by running
the script in \code{data-raw/maturity_at_age.R}.
}
\keyword{datasets}
