% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norpac_vesseltype.R
\name{norpac_vesseltype}
\alias{norpac_vesseltype}
\title{Change \code{VESSEL_TYPE} from numeric to character}
\usage{
norpac_vesseltype(x)
}
\arguments{
\item{x}{A vector of integers between one and six.}
}
\value{
A vector of combined integer values and character strings.
}
\description{

}
\author{
Kelli Faye Johnson
}
