% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norpaccatches.R
\name{norpaccatches}
\alias{norpaccatches}
\title{Workup NORPAC Catches}
\usage{
norpaccatches(ncatch = NULL, writecsv = TRUE, colour = TRUE, nyears = 5)
}
\arguments{
\item{ncatch}{An R object with NORPAC catches. If \code{NULL}, 
which is the default, then the \code{.Rdat} file will be read
from the disk.}

\item{writecsv}{A logical specifying if csv files should be
written to the disk.}

\item{colour}{A logical value specifying if the plots should be 
made in colour or gray scale, where the latter occurs if \code{FALSE}.}

\item{nyears}{The number of years for plotting.}

\item{x}{A named list of values by year, more than likely
made by tapply or split.}

\item{country}{The country the data is coming from, which
is only relevant if you are saving the data to the disk
for naming purposes.}

\item{type}{A character value that will be used for the name
of the file. 
The first portion should be the sector such as "atsea".
The second portion should be "fishing" if the file is not
bottom depth, which what the Canadians export.}

\item{dir}{A directory where you want to save the data. If left
\code{NULL} then no data will be saved and just a data frame will
be returned.}
}
\value{
Saves files to the disk
todo: summarize return in more detail
}
\description{
Summarize and plot NORPAC catches for the at-sea hake fishery.
Fleet types are assigned to create summaries by fleet and year.
}
\details{
*Tribal catches*
}
\author{
Kelli Faye Johnson
}
