% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pacfincatches.R
\name{pacfincatches}
\alias{pacfincatches}
\title{Clean catches extracted from PacFIN}
\usage{
pacfincatches(pcatch = NULL, addtribal = 0)
}
\arguments{
\item{pcatch}{An R object with PacFIN catches. The default value is
\code{NULL}, where the object will be read from a saved object within
a directory called extractedData.}

\item{addtribal}{A numeric value in metric tons that will be added
to the US shoreside catch to account for tribal fish tickets not
being in PacFIN yet.}
}
\value{
A list of three data frames,
\enumerate{
\item tribal catches,
\item removed foreign catches, and
\item catches by year and sector.
} are returned and the following files are saved to the disk:
\itemize{
\item PacFIN_Fleet.csv
\item PacFIN_Sector.cs
\item USshoreCatchByPeriodComp_ft.csv
\item us-shore-catch-by-month.csv
\item us-research-catch-by-month.csv
\item us-shore-startdate-by-dahl.csv
\item us-ti-catch-by-month.csv
}
}
\description{
Remove foreign catches and fix column names of catches from PacFIN.
}
