% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_boxplot.R
\name{plot_boxplot}
\alias{plot_boxplot}
\title{Plot a box plot with user-defined probabilities}
\usage{
plot_boxplot(
  data,
  xvar,
  yvar,
  file,
  probabilities = c(0.025, 0.25, 0.5, 0.75, 0.975),
  showmedian = FALSE,
  mlab = "Box plot with 95 percent quantiles",
  xlab = tools::toTitleCase(xvar[1]),
  ylab = "",
  yscale = "identity",
  incolor,
  scales = "fixed",
  nrow = NULL,
  ncol = NULL,
  legend.position,
  legend.direction = "vertical",
  ...
)
}
\arguments{
\item{data}{A data frame.}

\item{xvar}{A vector of column names available in \code{data}.}

\item{yvar}{A character value representing the column name
in \code{data} that is the dependent variable. There can only
be one entry in this argument, not a vector, and it must
match an existing column exactly.}

\item{file}{A file path for either an existing file or
a file that will be written to the disk by the given
function. Paths can be either full or relative.}

\item{probabilities}{A vector of five probabilities,
starting with the lower limit of the whisker,
then the lower edge of the box,
the middle of the box,
the upper edge of the box, and
the upper limit of the whisker.}

\item{showmedian}{A logical, specifying if a line should be plotted
for the third entry in \code{probabilities}, which will be the middle of
the box. The default, \code{FALSE}, leads to no median/middle line.}

\item{mlab}{The main title, located above the figure, which will
be centered. Please use sentence case without a full stop.
To create a new line use \verb{\\n} in the text.}

\item{xlab}{A label for the x axis. The default is to use the
x-axis variable in sentence case without a full stop.
Remember to give units if applicable.}

\item{ylab}{A label for the y axis. The default label is typically empty so
users are reminded to explain their y-axis label rather than just
using a column name. Please use sentence case without a full stop and
units should be given if applicable.}

\item{yscale}{The desired scale for the y axis. Allowable values
are all of those accepted in the \code{trans} argument of
\link[ggplot2:scale_continuous]{ggplot2::scale_y_continuous}. The default, \code{identity}, leads
to nothing being changed. Users will typically want to change
the value to \code{log}, which doesn't transform the data just the
distance between the labels on the y axis.}

\item{incolor}{If missing, no colors will be used for \code{aes(fill = )} in
the boxplots. Otherwise, a vector of column names can be used, which
will lead to \link{interaction} being called and the resulting factor
will determine the color categories. Colors will be determined
using \link{plot_colour}; but, note that this internal function can only
handle a limited number of categories.}

\item{scales}{A character value supplied to \link[ggplot2:facet_wrap]{ggplot2::facet_wrap} that
defines how the scales of the axes are determined across facets.
The default of \code{fixed} causes all axes to be the same across panels.
The most-likely value besides the default will be \code{free} allowing axes to
be specific to each panel.}

\item{nrow, ncol}{Number of rows and columns passed to
\link[ggplot2:facet_wrap]{ggplot2::facet_wrap}. The default values of \code{NULL} allow the function to
determine the appropriate number of rows and columns. If changed from the
default, use integer values to predetermine the number of rows or columns.}

\item{legend.position}{Coordinates for where you would like the legend
to be placed.}

\item{legend.direction}{\code{"vertical"} or `"horizontal",
where the default is the former.}

\item{...}{Arguments passed to \link[ggplot2:ggsave]{ggplot2::ggsave} other than \code{filename}
which is passed using \code{file}.}
}
\value{
A \code{ggplot2} object is invisibly returned. If \code{file} is supplied,
then the figure is also saved based on the entry (i.e., either relative
to your current working directory or based on the full path if specified).
The returned object contains a data frame of the data used in the figure
and not the data frame supplied to the figure. This object \code{returned[["data"]]}
is helpful when trying to determine quantiles for text in a document.
}
\description{
Plot a box plot with user-defined probabilities using
\link[ggplot2:geom_boxplot]{ggplot2::geom_boxplot} and \link[ggplot2:facet_wrap]{ggplot2::facet_wrap} if
more than one \code{xvar} is supplied.
}
\examples{
testdata <- data.frame(
  yvar = rnorm(100),
  xvar1 = rep(c("Good", "Bad", "Bad", "Good"), length.out = 100),
  xvar2 = rep(c("Apples", "Bananas", "Oranges", "Carrots"), length.out = 100)
)
gg <- plot_boxplot(
  data = testdata, xvar = c("xvar1"), yvar = "yvar",
  showmedian = TRUE, ylab = "Best fruit", incolor = "xvar2",
  legend.position = "right"
)
\dontrun{
print(gg)
grDevices::dev.off()
rm(testdata, gg)
}
}
