% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_rawmeasure.R
\name{plot_rawmeasure}
\alias{plot_rawmeasure}
\title{Workup At-Sea Composition Data for Pacific Hake}
\usage{
plot_rawmeasure(atsea.ages = NULL, years = 2017)
}
\arguments{
\item{atsea.ages}{An R object with NORPAC ages. The default value is
\code{NULL}, where the object will be read from a saved object within
a directory called extractedData.}

\item{years}{}

\item{ncatch}{An R object with NORPAC catches. The default value is
\code{NULL}, where the object will be read from a saved object within
a directory called extractedData.}
}
\value{
#todo: document return
}
\description{
todo: add more information about this function here.
}
\author{
Kelli Faye Johnson
}
