% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_weight_at_age.R
\name{plot_weight_at_age}
\alias{plot_weight_at_age}
\title{Plot weight-at-age data}
\usage{
plot_weight_at_age(data, maxage)
}
\arguments{
\item{data}{A data frame of weight-at-age values.}

\item{maxage}{The maximum age in the modelled data.}
}
\value{
A {ggplot2} object.
}
\description{
Plot weight-at-age data
}
\author{
Kelli F. Johnson
}
