% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_age-.R
\name{process_age_sea}
\alias{process_age_sea}
\title{Process age data from the Pacific Hake U.S. at-sea fishery}
\usage{
process_age_sea(
  atsea.ages = get_local(file = "atsea.ages.Rdat"),
  ncatch = get_local(file = "ncatch.Rdat"),
  years = 2008:hakedata_year(),
  ages = 1:15,
  files = fs::path(hakedata_wd(), paste0("us-", c("cp", "ms"), "-age-proportions.csv")),
  write = TRUE
)
}
\arguments{
\item{atsea.ages}{An R object with NORPAC ages. The default value is
\code{NULL}, where the object will be read from a saved object within
a directory called extractedData.}

\item{ncatch}{An R object of NORPAC catches.
The default value is loaded from the disk using \code{\link[=get_local]{get_local()}}.
Otherwise, the R object is typically stored in the \code{hakedata} environment.}

\item{years}{A vector of integers specifying the years of data that you wish
to process. Typically all years prior to 2008 are left static despite
updates to the data.}

\item{ages}{A vector of ages to be included in the composition data.
The default is to include ages from one to fifteen.}

\item{files}{File paths to the exported csv files. Must have lower case
\code{"cp"} or \code{"ms"} in the file name so the correct vessel type is chosen.}

\item{write}{A logical specifying if the data should be written to the disk.
The default is \code{TRUE}.}
}
\value{
A list object with summary information regrading the composition
data. Multiple files are saved to the disk as well. These saved files
include summary information in ...report.txt files and comps.csv files.
}
\description{
Process age data from the Pacific Hake U.S. at-sea fishery
}
\author{
Kelli F. Johnson
}
