% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_database.R
\name{process_database}
\alias{process_database}
\title{Process the raw data from \code{\link[=pull_database]{pull_database()}}}
\usage{
process_database()
}
\description{
Process the raw data from \code{\link[=pull_database]{pull_database()}}
}
\concept{process}
