% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_catch-.R
\name{process_vessel_type}
\alias{process_vessel_type}
\title{Change \code{VESSEL_TYPE} from numeric to character}
\usage{
process_vessel_type(x)
}
\arguments{
\item{x}{A vector of integers between one and six.}
}
\value{
A vector of combined integer values and character strings.
}
\description{
\code{VESSEL_TYPE} in the NORPAC data base are integers and this function
changes these integer values to character strings that provide meaning.
The following integers are the available options in NORPAC to indicate
whether the vessel processes fish or delivers it to a processing plant:
\itemize{
\item a catcher processor (CP) vessel,
\item a mothership (MS) or ship that receives unsorted codends from
other vessels,
\item a catcher only vessel that delivers unprocessed fish to shoreside,
floating plant, or vessel,
\item a MS that receives sorted codends,
\item a vessel that sells the majority of their catch over the side to other
fishing vessels who will utilize the fish for bait, and
\item vessels that discard all catch from a haul.
}
}
\author{
Kelli F. Johnson
}
