% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_weight_at_age-.R
\name{process_weight_at_age}
\alias{process_weight_at_age}
\title{Create weight-at-age files for hake assessment}
\usage{
process_weight_at_age(
  dir = hakedata_wd(),
  maxage = 15,
  max_year = hakedata_year(),
  navgyears = 5,
  nforecast = 4,
  maturity = maturity_at_age,
  output_wtatage_file_name = "wtatage.ss"
)
}
\arguments{
\item{dir}{The directory where the data is stored.
It can either be relative or absolute because the working
directory will not be changed. Instead, \code{dir} is
used to import data and export results to.}

\item{maxage}{The age of the plus group used for the stock assessment.
This will correspond to the maximum age group in the data, not in the
model because SS can model many ages when there is only information in
the data for a few ages.}

\item{max_year}{A four-digit integer specifying the maximum year of data
that you want to include in the weight-at-age data. The default is the
last year of data found using \code{\link[=hakedata_year]{hakedata_year()}}.}

\item{navgyears}{The number of early and late years to average since 1975 and
\code{max_year} for the early and late analysis asked for by the Scientific
Review Group in 2017. The argument can be a single value or a vector of two
values, where in the latter case the second value will be used for the most
recent time period.}

\item{nforecast}{The number of years to forecast into the future.
Typically, this is three for the hake assessment and will lead to
this many rows of mean weight-at-age data being copied to the data frame
beyond the last year of empirical data.}

\item{maturity}{A vector of maturity values from the maturity ogive. The
length needs to be the same as the number of ages in the model, not the
number of ages in the data. The default is to use the maturity ogive stored
in the package.}
}
\value{
todo: document return
}
\description{
Create weight-at-age files for the hake stock assessment.
Updated csv or rds files must exist prior to running.
}
\author{
Ian G. Taylor
}
