% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processNorpacCatch.R
\name{processNorpacCatch}
\alias{processNorpacCatch}
\title{Process Raw Data Extracted From NORPAC}
\usage{
processNorpacCatch(ncatch, species = 206, outfname = NULL, verbose = FALSE)
}
\arguments{
\item{ncatch}{An R object with NORPAC catches. The default value is
\code{NULL}, where the object will be read from a saved object within
a directory called extractedData.}

\item{species}{A numeric value that denotes the species of interest in the
column \code{SPECIES}.}

\item{outfname}{A file path to save the output to.}

\item{verbose}{A logical value specifying whether or not to print
information to the screen.}
}
\value{
\enumerate{
  \item A data frame of monthly catches in metric tons.
  \item A csv file is saved if \code{outfname} is specified.
}
}
\description{
Raw data consists of sampled and unsampled hauls with the total weight
including bycatch weight. For sampled hauls, \code{EXTRAPOLATED_WEIGHT}
is the weight of the catch only. For unsampled hauls that have no knowledge
of the amount of bycatch included in the haul, \code{OFFICIAL_TOTAL_CATCH}
must be used along with a estimated bycatch rate to calculate
\code{EXTRAPOLATED_WEIGHT}. Monthly bycatch rates are estimated from
the sampled hauls and multiplied by the \code{OFFICIAL_TOTAL_CATCH} to get
an estimate of the amount of bycatch that should be subtracted from the total
weight to get just the weight of the species of interest.
}
\details{
Definitions
\enumerate{
  \item @template official_total_catch
  \item @template extrapolated_weight
}

Assumptions
\enumerate{
  \item sampled hauls have an entry in the \code{SPECIES} column,
  i.e., hake are denoted with 206
  \item an unsampled haul is represented by one unique record
}
}
\author{
Andi Stephens
}
