% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_database.R
\name{pull_database}
\alias{pull_database}
\title{Pull data from databases for Pacific Hake}
\usage{
pull_database(
  database = c("NORPAC", "PacFIN"),
  startyear = list(NORPAC = 2008, PacFIN = c(1980, 2008)),
  endyear = hakedata_year(),
  passwordfile = "password.txt",
  savedir = hakedata_wd()
)
}
\arguments{
\item{database}{A vector of character strings indicating
which databases you want to pull information from.
Options include \code{c("NORPAC", "PacFIN")},
one or both (the default) can be specified.
Note that case matters here.}

\item{startyear}{An ordered list the same length as \code{database} with
at least one integer entry per list item specifying the
first year that you want data for.
If only a single value per database is provided, then
then the year will be recycled for catch and composition information.
Otherwise,
the first entry is the start year for catches and
the second is the start year for the biological data.
Names of the list must match entries in \code{database} or be \code{NULL}.}

\item{endyear}{An integer value specifying the final year of data
to include in the extraction. The default uses
\code{hakedata_year()} to determine the terminal year.}

\item{passwordfile}{A file path directing R to a file that contains
a password for each database listed in \code{database}.
Each password should be on its own line. The default case would place
your NORPAC password on the first line and the
second line would be your PacFIN password without quotes.
If this argument is \code{NULL}, users will be prompted
for their passwords.
Passwords are needed because the databases store confidential data.}

\item{savedir}{A string specifying the path of interest.}
}
\value{
An environment or \code{list} with several objects pulled from the
NORPAC and PacFIN databases. Several \code{Rdat} files are
saved to the disk in the extractedData folder,
one file for each object and a summary file.
}
\description{
Extraction of
\itemize{
\item catch data,
\item weight measurements,
\item length measurements,
\item age reads, and
\item management quantities
from NORPAC and PacFIN databases for the Pacific Hake assessment.
Other, additional, databases may be added in the future through the
\verb{database =} argument.
}
}
\details{
\code{pull_database()} includes several steps, outlined below,
and leads to many files being saved to your computer.
Originally, files were \strong{NOT} overwritten. In 2022, Kelli F. Johnson
changed the code to not save previous downloads to
save on disk space, reduce the storing of redundant information, and
simplify the code base that needed maintenance.
\itemize{
\item Find the folder where the data should be saved with \code{\link[=hakedata_wd]{hakedata_wd()}}.
\item Extract data from NORPAC.
\itemize{
\item Catch data
\item Weight and age data
\item Squash table of ages (that also includes lengths)
\item Foreign ages
\item Species list
}
\item Extract data from PacFIN.
\itemize{
\item Catch data
\item Age, length, and weight data from bds table
\item At-sea data
\item Management quantities such as allowable catch limits (ACL) for the
Pacific Fisheries Management Council, i.e., U.S. quantities of interest.
}
\item Save each object to the disk in the "extractedData" directory in
\code{\link[=hakedata_wd]{hakedata_wd()}}.
}
}
\examples{
\dontrun{
# An environment with objects is returned
dataenv <- pull_database()
# Access individual objects using `get()`
head(get("ncatch", envir = dataenv))
# Access individual objects pretending the environment is a list
dataenv[["ncatch"]][1:5, ]
}

}
\seealso{
\itemize{
\item \code{\link[=hakedata_year]{hakedata_year()}}
\item \code{\link[=hakedata_wd]{hakedata_wd()}}
}
}
\author{
Kelli F. Johnson
}
