% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pulldatabase.R
\name{pulldatabase}
\alias{pulldatabase}
\title{Pull Information From NORPAC and PacFIN Databases}
\usage{
pulldatabase(
  database = c("NORPAC", "PacFIN"),
  startyear = list(NORPAC = 2008, PacFIN = c(1980, 2008)),
  endyear = as.numeric(format(Sys.time(), "\%Y")),
  passwordfile = "password.txt"
)
}
\arguments{
\item{database}{A vector of character values indicating 
which databases you want to pull information from.
Options include \code{c("NORPAC", "PacFIN")}, one or both
can be specified.}

\item{startyear}{An ordered list the same length as \code{database} with
at least one element per list entry specifying the start year for
each database. If only a single value per database is provided, then
then the number will be recycled for catches and ages. Otherwise,
the first entry is the start year for catches and the second is the 
start year for age compositions.}

\item{endyear}{An integer value specifying the final year of data
to include in the extraction. The default will be to
end the data based on the current calendar year.}

\item{passwordfile}{A file path directing R to a file that contains
two lines, with the first being your NORPAC password and the
second being your PacFIN password without quotes.
If this argument is \code{NULL}, the user will be prompted
for their password.}
}
\value{
An environment with several objects pulled from the 
NORPAC and PacFIN databases. Several \code{Rdat} files are
saved to the disk, i.e., one for each object and a summary 
file.
}
\description{
Extract catch and age data from NORPAC and PacFIN databases
for the hake stock assessment.
}
\examples{
dataenv <- pulldatabase()
head(get("ncatch", envir = dataenv))

}
\author{
Kelli Faye Johnson
}
