% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shorecomps.R
\name{shorecomps}
\alias{shorecomps}
\title{Workup Shoreside Compsition Data}
\usage{
shorecomps(page = NULL, ages = 1:15, verbose = FALSE)
}
\arguments{
\item{page}{PacFIN biological data from the comprehensive database.}

\item{ages}{Vector of ages to keep.}

\item{verbose}{A logical specifying whether or not to print
information to the screen.
The default is \code{FALSE}.}
}
\value{
todo: document the return
}
\description{
Provide composition data from the U.S. Shoreside
hake fishery. Written by Allan Hicks in 2016 and
revised in 2017 by Ian Taylor.
}
\author{
Kelli Faye Johnson
}
