% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_weight_at_age.R
\name{test_weight_at_age}
\alias{test_weight_at_age}
\title{Run tests for historical weight-at-age data}
\usage{
test_weight_at_age(data)
}
\arguments{
\item{data}{A data frame of weight-at-age data.}
}
\value{
A logical of \code{TRUE} if it reaches the end without error.
}
\description{
Run tests for historical weight-at-age data
}
\seealso{
Other weight-at-age: 
\code{\link{weight_at_age_outlier}()},
\code{\link{weight_at_age_read}()}
}
\author{
Kelli F. Johnson
}
\concept{weight-at-age}
