% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UScatchPlots.R
\name{UScatchPlots}
\alias{UScatchPlots}
\title{Plots of Catch for US Only}
\usage{
UScatchPlots(doPNG = TRUE, nyears = 5)
}
\arguments{
\item{doPNG}{A logical specifying if png files should be saved.}

\item{nyears}{The number of years you want to plot.}
}
\value{
todo: document return of UScatchPlots
}
\description{
Plots for the JTC of US catch data.
todo: Make CP and MS use NORPAC data instead of PacFIN.
}
\author{
Kelli Faye Johnson
}
