% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workupNorpacCatchSpecies.R
\name{workupNorpacCatchSpecies}
\alias{workupNorpacCatchSpecies}
\title{Summarize Bycatch in the Pacific Hake Fishery With Barplots}
\usage{
workupNorpacCatchSpecies(
  ncatch = NULL,
  nspecies = NULL,
  fileout = "bycatchsummary.pdf"
)
}
\arguments{
\item{ncatch}{An R object with NORPAC catches. The default value is
\code{NULL}, where the object will be read from a saved object within
a directory called extractedData.}

\item{nspecies}{A data frame of species names from NORPAC.}

\item{fileout}{A file path to a \code{.pdf} extension to specify where
the plots should be saved. A single \code{.pdf} can be created or multiple
\code{.png} files will be created.}
}
\value{
A data frame of bycatch by species, year, and sector.
}
\description{
Create barplots of bycatch by species for the Pacific hake fishery.
Data is summarized from the NORPAC database only.
}
\author{
Kelli Faye Johnson
}
