% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wtatage_calculations.R
\name{write_wtatage_file}
\alias{write_wtatage_file}
\title{Write weight-at-age file}
\usage{
write_wtatage_file(
  file = paste0("wtatage_", format(Sys.time(), "\%Y"), "created_", format(Sys.time(),
    "\%d-\%b-\%Y_\%H.\%M"), ".ss"),
  data,
  verbose = FALSE
)
}
\arguments{
\item{file}{A filename that you want to save the information as. The default is to have
an extension of \code{.ss} such that the file can be used for Stock Synthesis.
The file path can either be relative or absolute.}

\item{verbose}{A logical value specifying whether or not to print
information to the screen.}
}
\description{
Write weight-at-age file
}
