% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wtatage_extra.R
\name{wtatage_extra}
\alias{wtatage_extra}
\title{todo: give a better description}
\usage{
wtatage_extra(
  dir,
  outliers = TRUE,
  maxage = 15,
  yrs = 2008:format(Sys.Date(), "\%Y"),
  navgyears = c(5, 3)
)
}
\arguments{
\item{dir}{The directory where the data is stored. 
It can either be relative or absolute, but no working
directory will be changed. Instead, the \code{dir} is
just used to import data and save resulting plots in.}

\item{outliers}{A logical value if outliers should be looked for
in the individual files or if a summary file that was previously 
generated using \code{weightatage} should be loaded.}

\item{maxage}{The age of the plus group used for the stock assessment.}

\item{yrs}{A vector of years to search for recent data. Typically,
the vector starts with 2008 and ends with the most recent year
of data.}

\item{navgyears}{The number of early and late years to average since
1975 and \code{max(yrs)} for the early and late analysis asked for
by the Scientific Review Group in 2017. The argument can be a single
value or a vector of two values, where in the latter case the second
value will be used for the most recent time period.}
}
\value{
todo: document return
}
\description{
todo: give a better description
}
\examples{
weightatage(dir = file.path(hakedatawd(), "LengthWeightAge"), yrs = 2008:2017)

}
\author{
Ian Taylor
}
