% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatoassessment_comp.R
\name{datatoassessment_comp}
\alias{datatoassessment_comp}
\title{Copy composition data in the assessment data directory}
\usage{
datatoassessment_comp(dirout)
}
\arguments{
\item{dirout}{A directory that houses the \code{.csv} data files
for the stock assessment.}
}
\description{
Copy composition data in the assessment data directory
}
