% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hakedata_globals.R
\name{hakedatasqlpw}
\alias{hakedatasqlpw}
\title{Find username and passwords for databases}
\usage{
hakedatasqlpw(database = c("NORPAC", "PacFIN"), file)
}
\arguments{
\item{database}{A vector of character values specifying which databases you
want login information for.}

\item{file}{A file path specifying where to find the passwords.
The path can be full or relative to your current working directory.
If a path is provided, the file that it leads to
must be for a text file with one password per line for each database
in the \code{database} argument and in that order.
The default for \code{database} means that the file would have two lines,
where the first line is the NORPAC password and
the second line is the PacFIN password.
These passwords should not be surrounded with quotes.
If a file name is not provided, which is the default behaviour, then
the user will be prompted for their passwords. This also happens if
the file cannot be found given the path provided.}
}
\value{
A list with two entries, \code{usernames} and \code{passwords}.
Each element will have the same number of entries as the
input argument \code{database} and be named using the elements of \code{database}.
The list is invisibly returned to ensure that the passwords are not printed
to the screen. Thus, the function call should be assigned to an object.
}
\description{
Find the username and passwords specific given the username of the computer
and stored or entered passwords for accessing databases that store
confidential information about landings of Pacific Hake.
}
\examples{
\dontrun{
# Prompted for passwords for each database
test <- hakedatasqlpw()
# Prompted for passwords for each database because file is not found
test <- hakedatasqlpw(file = "doesnotwork.txt")
# On Kelli Johnson's machine, the following will work
test <- hakedatasqlpw(file = "password.txt")
# Doesn't work because entry for database is not in the list
# of allowed databases, i.e., the default for `database`.
test <- hakedatasqlpw(database = "onedatabase")
# Only look for one password
test <- hakedatasqlpw(database = "NORPAC")
}
}
\author{
Kelli F. Johnson
}
