% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hakedata_globals.R
\name{hakedatawd}
\alias{hakedatawd}
\title{Find the Working Directory for hake-data}
\usage{
hakedatawd()
}
\value{
A full file path, not a relative file path.
}
\description{
Find the working directory based on your user name
for the given computer that you on. The working
directory is for the hake-data repository cloned
from github.
}
\examples{
hakedatawd()

}
\author{
Kelli Faye Johnson
}
