% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_catch.R
\name{new_catch}
\alias{new_catch}
\title{Move US Catch Information to Assessment Files}
\usage{
new_catch(dirout, year, filedat = NULL)
}
\arguments{
\item{dirout}{A directory that houses the \code{.csv} data files
for the stock assessment.}

\item{year}{The year of data that you are particularly interested
in updating, although most years get updated since 2008.}

\item{filedat}{The Stock Synthesis dat file that you want to
integrate catches into. The default is \code{NULL}, which allows
you to run the function without writing a new dat file.}
}
\value{
todo: document what this function returns.
}
\description{
Integrate the latest catch information from the US hake fishery
into the hake-assessment files used in the building of the
stock assessment.
}
\author{
Kelli Faye Johnson
}
