% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norpaccatches.R
\name{norpaccatches}
\alias{norpaccatches}
\title{Workup NORPAC Catches}
\usage{
norpaccatches(
  ncatch = loadlocal(file = "norpac_catch.Rdat"),
  nyears = 5,
  species = 206,
  savedir = hakedatawd()
)
}
\arguments{
\item{ncatch}{An R object with catch information extracted from the
NORPAC database. The default is to read the \code{.Rdat} file
from the disk using \code{\link[=loadlocal]{loadlocal()}}. Else, users can just pass the R object
itself, which would typically be \code{hakedata[["ncatch"]]}.}

\item{nyears}{The number of years for plotting.}

\item{species}{An integer specifying the species for given data base.
For NORPAC data, this will be three digits and
is associated with the species name in
\code{NORPAC.DOMESTIC_SPECIES}, using the column \code{SPECIES_NO}. Note that this
value is often called the \code{OBS} species code in observer landings from the
Alaska Fisheries Science Center and is not the \code{AKR} species code.
For example, \code{206} is Pacific Hake and is the default.
For PacFIN data, this will be the entry in \code{SPID}.}
}
\value{
Figures and csv files are saved to the disk regarding
catch rates and depth (fishing and bottom depths).
In the \code{Figures} directory, the following png or csv files are saved:
\itemize{
\item fishDepthsUS.png
\item fishDepthsUSallyears.png
\item fish_FISHING_DEPTH_FATHOMS_US.png
\item fish_BOTTOM_DEPTH_FATHOMS_US.png
\item fishBottomDepthByVesselUS.png
\item fishCatchRatesUSByYear.png
\item fishDepthByYearUS.png
\item fishCatchRatesUS.png
\item fishCatchRatesUSnolog.png
\item NORPAC_DomesticAtSea_bdepthfathom_range.csv
\item NORPAC_DomesticAtSea_bdepthfathom_deeper1500f.csv
\item us-cp-startdate.csv
\item us-ms-startdate.csv
}

In the \code{Catches} directory, the following csv files are saved:
\itemize{
\item depth-us-atsea-bottom.csv
\item depth-us-atsea-fishing.csv
}
}
\description{
Summarize and plot NORPAC catches for the at-sea hake fishery.
Fleet types are assigned to create summaries by fleet and year.
}
\details{
\emph{Tribal catches}
}
\section{Raw data}{
Raw data consists of sampled and unsampled hauls with the total weight
including bycatch weight. For sampled hauls, \code{EXTRAPOLATED_WEIGHT}
is the weight of the catch only. For unsampled hauls that have no knowledge
of the amount of bycatch included in the haul, \code{OFFICIAL_TOTAL_CATCH}
must be used along with a estimated bycatch rate to calculate
\code{EXTRAPOLATED_WEIGHT}. Monthly bycatch rates are estimated from
the sampled hauls and multiplied by the \code{OFFICIAL_TOTAL_CATCH} to get
an estimate of the amount of bycatch that should be subtracted from the total
weight to get just the weight of the species of interest.
}

\section{Definitions}{
\enumerate{
\item @template official_total_catch
\item @template extrapolated_weight
}
}

\section{Assumptions}{
\enumerate{
\item sampled hauls have an entry in the \code{SPECIES} column,
i.e., hake are denoted with 206
\item an unsampled haul is represented by one unique record
}
}

\author{
Kelli F. Johnson
}
