% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pacfincatches.R
\name{pacfincatches}
\alias{pacfincatches}
\title{Print summaries and figures for catches from PacFIN}
\usage{
pacfincatches(
  pcatch = loadlocal(file = "pacfin_catch.Rdat"),
  nyears = 5,
  savedir = hakedatawd()
)
}
\arguments{
\item{pcatch}{An R object of PacFIN catches.
The default value is loaded from the disk using \code{\link[=loadlocal]{loadlocal()}}.
Otherwise, the R object is typically stored in the \code{hakedata} environment.}

\item{nyears}{An integer specifying the number of years you want to plot,
where the most recent year will be sequenced back in time \code{nyears}.
The default is typically five years.}

\item{savedir}{A full path to a directory where you want files saved.
The default is to save them in a folder found using \code{\link[=hakedatawd]{hakedatawd()}},
which returns a path specific to the personal preferences of JTC members.}
}
\value{
The following files are saved to the disk:
\itemize{
\item PacFIN_Fleet.csv
\item PacFIN_Sector.cs
\item USshoreCatchByPeriodComp_ft.csv
\item us-shore-catch-by-month.csv
\item us-research-catch-by-month.csv
\item us-shore-startdate-by-dahl.csv
\item us-ti-catch-by-month.csv
}
}
\description{
Print summaries and figures for catches from PacFIN
}
