% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pacfincatches.R
\name{pacfincatches}
\alias{pacfincatches}
\title{Clean catches extracted from PacFIN}
\usage{
pacfincatches(pcatch = NULL)
}
\arguments{
\item{pcatch}{An R object with PacFIN catches. The default value is
\code{NULL}, where the object will be read from a saved object within
a directory called extractedData.}
}
\value{
A list of three data frames,
\enumerate{
  \item tribal catches,
  \item removed foreign catches, and
  \item catches by year and sector.
}
}
\description{
Remove foreign catches and fix column names of catches from PacFIN.
}
