% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_catchvmonthbyyear.R
\name{plot_catchvmonthbyyear}
\alias{plot_catchvmonthbyyear}
\title{Plot catch versus month by year}
\usage{
plot_catchvmonthbyyear(
  data,
  file,
  title,
  Yrs,
  quotas = NULL,
  cols,
  leg.cex = 1.2,
  divisor = 1000
)
}
\arguments{
\item{data}{A data frame.}

\item{file}{todo: document}

\item{title}{todo: document}

\item{Yrs}{todo: document}

\item{quotas}{todo: document}

\item{cols}{todo: document}

\item{leg.cex}{todo: document}

\item{divisor}{todo: document}
}
\description{
todo:document plot_catchvmonthbyyear
}
