% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_rawmeasure.R
\name{plot_rawmeasure}
\alias{plot_rawmeasure}
\title{Workup At-Sea and PacFIN lengths and weights}
\usage{
plot_rawmeasure(
  atsea.ages = NULL,
  page = NULL,
  years = as.numeric(format(Sys.time(), "\%Y")) - 0:4,
  xlim_weight = c(0, 3),
  xlim_length = c(0, 70)
)
}
\arguments{
\item{atsea.ages}{An R object with NORPAC ages. The default value is
\code{NULL}, where the object will be read from a saved object within
a directory called extractedData.}

\item{years}{A vector of years that you want plotted.
The default is to do the last five years.}

\item{ncatch}{An R object of NORPAC catches.
The default value is loaded from the disk using \code{\link[=loadlocal]{loadlocal()}}.
Otherwise, the R object is typically stored in the \code{hakedata} environment.}
}
\value{
Four figures are saved to the disk. Figures are of length
and weight distributions for the at-sea and shoreside sectors.
}
\description{
Workup At-Sea and PacFIN lengths and weights
}
\author{
Kelli Faye Johnson
}
