% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_weight_at_age-.R
\name{process_weight_at_age_US}
\alias{process_weight_at_age_US}
\title{Collate some of the age-weight data for the most recent year}
\usage{
process_weight_at_age_US(year = hakedata_year(), savedir = hakedata_wd())
}
\arguments{
\item{year}{An integer value specifying the year of data you would like to summarize.}
}
\value{
A list of weight-at-age information that includes the following items:
\enumerate{
\item acousticmean: Mean weight-at-age from the survey, if the s
survey operated in that year.
\item usatseamean: mean weight-at-age from the US-at-sea data.
\item usshorebasedmean: mean weight-at-age from the US-shore-based data.
\item outliers: A table of outliers by Agency for the US-shore-based fishery.
\item largefishtable: A table of large fish (i.e., greater than 10 kg)
by data source.
\item dat: The data that was saved to the disk for the given year.
\item survey: A logical value specifying if it was a survey year.
\item year: The year for which the data was compiled.
\item file: The file path used to save the \code{dat}.
}
}
\description{
This script contains commands to format the data on weight, length, and age
from different sources, including the acoustic survey, the U.S. at-sea fishery,
and the U.S. shoreside fishery. Canadian fishery data does not appear to be included.
}
\details{
The resulting compiled data for the most recent year are written to a file
with a name like /LengthWeightAge/LWAdata_2017.csv.

TODO: Ensure that this is done from 2008 to present every year to
because there may be changes to previous weight-at-age data.
}
\author{
Ian G. Taylor
}
