% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pulldatabase.R
\name{pulldatabase}
\alias{pulldatabase}
\title{Pull data from databases for Pacific Hake}
\usage{
pulldatabase(
  database = c("NORPAC", "PacFIN"),
  startyear = list(NORPAC = 2008, PacFIN = c(1980, 2008)),
  endyear = hakedata_year(),
  passwordfile = "password.txt"
)
}
\arguments{
\item{database}{A vector of character values indicating
which databases you want to pull information from.
Options include \code{c("NORPAC", "PacFIN")}, one or both
can be specified. The default is to pull from both data bases.}

\item{startyear}{An ordered list the same length as \code{database} with
at least one element per list entry specifying the start year for
each database. If only a single value per database is provided, then
then the number will be recycled for catches and ages. Otherwise,
the first entry is the start year for catches and the second is the
start year for the biological data.}

\item{endyear}{An integer value specifying the final year of data
to include in the extraction. The default uses
\code{hakedata_year} to determine the terminal year.}

\item{passwordfile}{A file path directing R to a file that contains
a password for each database listed in \code{database}.
Each password should be on its own line. The default case would place
your NORPAC password on the first line and the
second line would be your PacFIN password without quotes.
If this argument is \code{NULL}, users will be prompted
for their passwords.
Passwords are needed because the databases store confidential data.}
}
\value{
An environment or \code{list} with several objects pulled from the
NORPAC and PacFIN databases. Several \code{Rdat} files are
saved to the disk in the extractedData folder,
one file for each object and a summary file.
}
\description{
Extraction of
\itemize{
\item catch data,
\item weight measurements,
\item length measurements,
\item age reads, and
\item management quantities
from databases for the Pacific Hake assessment.
Current databases that are queried include the NORPAC and PacFIN databases,
though other databases may be added in the future through the
\verb{database =} argument.
}
}
\details{
There are many detailed parts to this function that lead to files
being saved to the disk. The steps are outlined below:
\itemize{
\item \code{\link[=hakedatawd]{hakedatawd()}} finds the folder where the data should be saved.
\item Extract data from NORPAC.
\itemize{
\item Catch data
\item Weight and age data
\item Squash table of ages (that also includes lengths)
\item Foreign ages
\item Species list
}
\item Extract data from PacFIN.
\itemize{
\item Catch data
\item Age, length, and weight data from bds table
\item At-sea data
\item Management quantities such as allowable catch limits (ACL) for the
Pacific Fisheries Management Council, i.e., U.S. quantities of interest.
}
\item Save each object to the disk in the "extractedData" directory in
\code{\link[=hakedatawd]{hakedatawd()}}.
}
}
\examples{
\dontrun{
# An environment with objects is returned
dataenv <- pulldatabase()
# Access individual objects using get
head(get("ncatch", envir = dataenv))
# Access individual objects pretending the environment is a list
dataenv[["ncatch"]][1:5, ]
}

}
\author{
Kelli Faye Johnson
}
