% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pulldatabase.R
\name{pulldatabase}
\alias{pulldatabase}
\title{Pull Information From NORPAC and PacFIN Databases}
\usage{
pulldatabase(
  database = c("NORPAC", "PacFIN"),
  startyear = list(NORPAC = 2008, PacFIN = c(1980, 2008)),
  endyear = hakedata_year(),
  passwordfile = "password.txt"
)
}
\arguments{
\item{database}{A vector of character values indicating 
which databases you want to pull information from.
Options include \code{c("NORPAC", "PacFIN")}, one or both
can be specified.}

\item{startyear}{An ordered list the same length as \code{database} with
at least one element per list entry specifying the start year for
each database. If only a single value per database is provided, then
then the number will be recycled for catches and ages. Otherwise,
the first entry is the start year for catches and the second is the 
start year for age compositions.}

\item{endyear}{An integer value specifying the final year of data
to include in the extraction. The default uses
\code{\link{hakedata_year}} to determine the terminal year.}

\item{passwordfile}{A file path directing R to a file that contains
two lines, with the first being your NORPAC password and the
second being your PacFIN password without quotes.
If this argument is \code{NULL}, the user will be prompted
for their password.}
}
\value{
An environment with several objects pulled from the 
NORPAC and PacFIN databases. Several \code{Rdat} files are
saved to the disk, i.e., one for each object and a summary 
file.
}
\description{
Extract catch, weight and age data from NORPAC and PacFIN databases
for the hake stock assessment.
}
\details{
There are many detailed parts to \code{pulldatabase} that lead to many files
being saved to the disk. The steps are outlined below:
\enumerate{
  \item The folder where the data should be saved is found.
  \item A summary file is saved to the disk.
  \item Extract data from NORPAC
    \enumerate{
      \item Catch data
      \item Weight and age data
      \item Squash table of ages (that also includes lengths)
    }
  \item Extract data from PacFIN
    \enumerate{
      \item Catch data
      \item Age data from bds table
      \item Species cluster
      \item Bds fish table
      \item Bds species cluster
      \item At sea sector data
    }
  \item Save each item to the disk in the "extractedData" folder.
}
}
\examples{
\dontrun{
dataenv <- pulldatabase()
head(get("ncatch", envir = dataenv))
}

}
\author{
Kelli Faye Johnson
}
