% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_ss3_catch.R
\name{update_ss3_catch}
\alias{update_ss3_catch}
\title{Update the catches in a Stock Synthesis file}
\usage{
update_ss3_catch(dir_data, file_ss3)
}
\arguments{
\item{dir_data}{A path to where the input files are saved.}

\item{file_ss3}{A file path to the Stock Synthesis data file that you want to
integrate catches into.}
}
\value{
A Stock Synthesis data file is augmented and overwritten.
}
\description{
Update the catches in a Stock Synthesis file
}
\author{
Kelli F. Johnson
}
